ARM: dts: bcm283x: Convert to generic bias pinctrl
Since the pinctrl-bcm2835 support generic bias settings, we can
use the generic DT properties instead of vendor specific ones.
Because the bias settings always applies to the whole node, we need to
re-arrange the pins (inspired by mt7623.dtsi). While we are at this make
this consistent above all UART and MMC pins. The biggest benefit is
that we don't need to introduce some mapping from BCM2835 bias to BCM7211.

Signed-off-by: Stefan Wahren <wahrenst@gmx.net>